+++ nbgitpuller_link/cli.py
@@ -24,11 +24,18 @@ from .link import Link
 )
 @click.option(
     "--launch-path",
-    help="Relative path to file or directory in source repository to launch on target JupyterHub",
+    help="Relative path to file or directory in source repository to launch on target JupyterHub.",
     default="",
     show_default=True,
 )
-def main(jupyterhub_url, repository_url, branch, launch_path):
+@click.option(
+    "--interface",
+    type=click.Choice(["notebook", "lab"], case_sensitive=False),
+    default="notebook",
+    help="Open with classic Jupyter Notebook interface or next-generation JupyterLab.",
+    show_default=True,
+)
+def main(jupyterhub_url, repository_url, branch, launch_path, interface):
     """Generate an nbgitpuller link to load a repository on a JupyterHub"""
-    linker = Link(jupyterhub_url, repository_url, branch, launch_path)
+    linker = Link(jupyterhub_url, repository_url, branch, launch_path, interface)
     click.secho(linker.link, fg="green")