+++ CHANGES.md
@@ -4,7 +4,8 @@ Changes: nbgitpuller-link
 0.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+* Add choice of classic Notebook or JupyterLab interface
+* Test against known link
 
 
 0.1 (2021-02-28)